Review commit
Commit addressing review comments
